+++ server/src/main/kotlin/com/tomaszezula/makker/server/model/Requests.kt
@@ -5,17 +5,17 @@ import kotlinx.serialization.Serializable
 sealed interface Request
 
 @Serializable
-data class CreateScenarioRequest(val teamId: Long, val folderId: Long, val blueprint: String) : Request
+data class CreateScenarioRequest(val teamId: Int, val folderId: Int, val blueprint: String) : Request
 
 @Serializable
-data class UpdateScenarioRequest(val scenarioId: Long? = null, val blueprint: String) : Request
+data class UpdateScenarioRequest(val scenarioId: Int? = null, val blueprint: String) : Request
 
 @Serializable
-data class GetBlueprintRequest(val scenarioId: Long) : Request
+data class GetBlueprintRequest(val scenarioId: Int) : Request
 
 @Serializable
-data class SetModuleDataRequest(val scenarioId: Long? = null, val modules: List<Module>) :
+data class SetModuleDataRequest(val scenarioId: Int? = null, val modules: List<Module>) :
     Request {
     @Serializable
-    data class Module(val moduleId: Long, val key: String, val value: String)
+    data class Module(val moduleId: Int, val key: String, val value: String)
 }
